requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
passlib[bcrypt]
jose
python-dotenv
google-cloud-aiplatform
elasticsearch
psycopg2
pyjwt
sqlalchemy-utils
pypdf
langchain
pytest
boto3
langchain-pinecone
langchain-google-genai
sentence-transformers
langchain-huggingface
langchain-community
langchain-openai
unstructured
python-multipart 
langchain-chroma 
chromadb 
chromadb-client